1.0 - Only 1 case

+++ src/equationsystemproject/AnaliticMethod.java
@@ -4,7 +4,7 @@ public class AnaliticMethod {
 
 	double hx;
 
-	public double calculate(double l, int hl, int h0) {
+	public double calculate(double l, double hl, double h0) {
 		// hl/2=a1*(L/2)+h0
 		// a1=(1/L)*(hL-h0)
 		// a0=h0
